tts-service/requirements_tts.txt
# requirements_tts.txt

flask
flask-cors
redis
rq
paramiko
pydub
imageio-ffmpeg
piper-tts
gunicorn